fix: Consolidate authentication utilities into `shared.utils` and update imports across lambda functions.
