+++ shared_resources/python-modules/python/shared/utils/__init__.py
@@ -18,3 +18,4 @@ from .reference_utils import (
     update_references_table,
 )
 from .cognito_utils import get_cognito_user_by_id
+from .auth import require_permission, PermissionError, get_permissions_from_event
